# TinadecCore/tests/__snapshots__/openapi.core.json
{
  "components": {
    "schemas": {
...
      "ApprovalRuleCreateRequest": {
        "properties": {
          "kind": {
            "type": "string"
          },
          "pattern": {
            "type": [
              "null",
              "string"
            ]
          },
          "session_id": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "tool_id": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "tool_id"
        ],
        "type": "object"
      },
      "ApprovalRuleDto": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "created_by_principal_id": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "pattern": {
            "type": [
              "null",
              "string"
            ]
          },
          "session_id": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "tool_id": {
            "type": "string"
          },
          "use_count": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "tool_id",
          "pattern",
          "session_id",
          "created_by_principal_id",
          "created_at",
          "use_count"
        ],
        "type": "object"
      },
...
      "MessageAttachmentDto": {
        "properties": {
          "bound_at": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "content_hash": {
            "type": "string"
          },
          "content_length": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "file_name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "media_type": {
            "type": "string"
          },
          "message_id": {
            "type": [
              "null",
              "string"
            ]
          },
          "session_id": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TinadecCore.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/approval-rules": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRuleDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "ApprovalRules"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRuleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRuleDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "ApprovalRules"
        ]
      }
    },
    "/api/v1/approval-rules/{ruleId}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "ruleId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "ApprovalRules"
        ]
      }
    },
    "/api/v1/approvals": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
...
    "/api/v1/approvals/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "TinadecCore.Api"
        ]
      }
    },
...
    "/api/v1/attachments/{attachmentId}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Attachments"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageAttachmentDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Attachments"
        ]
      }
    },
    "/api/v1/attachments/{attachmentId}/content": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Attachments"
        ]
      }
    },
...
    "/api/v1/sessions/{sessionId}/attachments": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageAttachmentDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Attachments"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageAttachmentDto"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Payload Too Large"
          }
        },
        "tags": [
          "Attachments"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}